scmi: Introduce pinctrl scmi protocol driver
Implementation of the SCMI client driver, which implements
PINCTRL_PROTOCOL. This protocol has id 18 and described in
DEN0056C document.
This protocol is the part of the feature that was designed
to separate pinctrl subsystem to SCP firmware. The idea is to
separate communication of the pin control subsystemd with the HW
to SCP firmware (or similar system, such as ATF), which provides
interface to give OS ability to control HW through SCMI protocol.
This is generic driver, which implements SCMI protocol, independent
from the platform type.

DEN0056C document:
https://developer.arm.com/documentation/den0056/latest

Signed-off-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>