pinctrl: Implementation of the generic scmi-pinctrl driver
scmi-pinctrl driver implements pinctrl driver interface and using
SCMI protocol to redirect messages from pinctrl subsystem SDK to
SCP firmware, which does the changes in HW.

This setup expects SCP firmware (or similar system, such as ATF)
to be installed on the platform, which implements pinctrl driver
for the specific platform.

SCMI-Pinctrl driver should be configured from the device-tree.
The following binding shows the configuration for Renesas H3ULCB
for r8a77951 as an example:
 Documentation/devicetree/bindings/renesas-scmi,pfc.yaml

Signed-off-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>